feat(journey): push to db journey informations
