+++ app/store/services/users.service.ts
@@ -3,6 +3,11 @@ import { getAccessToken } from '../../utils/auth'
 const API_URL = 'https://viva-api.fly.dev'
 const ENDPOINT = '/user'
 
+export interface IInputJourney {
+  journey: string
+  score: number
+}
+
 export const changeCVService = async (cv: string) => {
   const accessToken = await getAccessToken()
   fetch(`${API_URL}${ENDPOINT}`, {
@@ -17,6 +22,30 @@ export const changeCVService = async (cv: string) => {
     .then(response => {
       if (response.data) {
         return response.data
+      }
+      return Promise.reject()
+    })
+    .catch(error => {
+      console.log('error', error)
+      return Promise.reject(error)
+    })
+}
+
+export const addJourneyService = async (journey: IInputJourney) => {
+  const accessToken = await getAccessToken()
+
+  fetch(`${API_URL}${ENDPOINT}`, {
+    method: 'PATCH',
+    headers: {
+      'Content-Type': 'application/json',
+      Authorization: `Bearer ${accessToken}`,
+    },
+    body: JSON.stringify({ journeys: [journey] }),
+  })
+    .then(response => response.json())
+    .then(response => {
+      if (response) {
+        return response
       }
       return Promise.reject()
     })
